fix: stop log to sentinel from triggering Slack retries
Update the `log_to_sentinel` method so that on failure, it logs
an error but does not bubble the exception up to the caller.
This is being done because a failure to log to Sentinel should not
break the SRE Bot's behaviour.

The logged error will get caught by the CloudWatch metric filter
and allow the SRE team to investigate.